resolve operational safety review findings
