+++ apps/web/package.json
@@ -8,6 +8,7 @@
     "start": "next start"
   },
   "dependencies": {
+    "@x-harness/content-os": "workspace:*",
     "next": "^15.5.21",
     "react": "^19.0.0",
     "react-dom": "^19.0.0"
