+++ apps/web/src/lib/cubelic-time.ts
@@ -1,3 +1,5 @@
+import { isValidCalendarDateTime } from '@x-harness/content-os';
+
 export function tokyoDateTimeLocalToIso(value: string): string {
   const match = value.match(/^(\d{4})-(\d{2})-(\d{2})T(\d{2}):(\d{2})(?::(\d{2}))?$/);
   if (!match) throw new Error('Asia/Tokyoの有効な日時を入力してください');
@@ -8,17 +10,7 @@ export function tokyoDateTimeLocalToIso(value: string): string {
   const hour = Number(hourText);
   const minute = Number(minuteText);
   const second = Number(secondText);
-  const leapYear = year % 4 === 0 && (year % 100 !== 0 || year % 400 === 0);
-  const daysInMonth = [31, leapYear ? 29 : 28, 31, 30, 31, 30, 31, 31, 30, 31, 30, 31][month - 1] ?? 0;
-  if (
-    month < 1
-    || month > 12
-    || day < 1
-    || day > daysInMonth
-    || hour > 23
-    || minute > 59
-    || second > 59
-  ) {
+  if (!isValidCalendarDateTime({ year, month, day, hour, minute, second })) {
     throw new Error('Asia/Tokyoの有効な日時を入力してください');
   }
   const instant = Date.parse(
