+++ packages/content-os/src/datetime.ts
@@ -1,0 +1,23 @@
+export interface CalendarDateTimeParts {
+  year: number;
+  month: number;
+  day: number;
+  hour: number;
+  minute: number;
+  second: number;
+}
+
+export function isValidCalendarDateTime(parts: CalendarDateTimeParts): boolean {
+  const { year, month, day, hour, minute, second } = parts;
+  const leapYear = year % 4 === 0 && (year % 100 !== 0 || year % 400 === 0);
+  const daysInMonth = [31, leapYear ? 29 : 28, 31, 30, 31, 30, 31, 31, 30, 31, 30, 31][month - 1] ?? 0;
+  return (
+    month >= 1
+    && month <= 12
+    && day >= 1
+    && day <= daysInMonth
+    && hour <= 23
+    && minute <= 59
+    && second <= 59
+  );
+}
