+++ packages/content-os/src/index.ts
@@ -1,4 +1,5 @@
 export * from './adapter.js';
+export * from './datetime.js';
 export * from './errors.js';
 export * from './generator.js';
 export * from './manual-authority.js';
